Check the downloadedDatabaseFile is not null before deleting
On the first sync (or when the db on the server is deleted) the downloadedDatabaseFile will be null

+++ src/com/u17od/upm/FullAccountList.java
@@ -375,7 +375,9 @@ public class FullAccountList extends AccountsList {
             UIUtilities.showToast(this, R.string.db_uptodate);
         }
         app.setTimeOfLastSync(new Date());
-        downloadedDatabaseFile.delete();
+        if (downloadedDatabaseFile != null) {
+            downloadedDatabaseFile.delete();
+        }
     }
 
     private void launchDonatePage() {
